learning/requirements.txt
PyYAML>=6.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
python-dotenv>=0.21.0